requirements.txt
streamlit==1.51.0
pillow
requests